+++ src/optimise/mod.rs
@@ -1,6 +1,4 @@
 mod anneal;
-pub use anneal::{anneal_enumerations, template_weight};
-mod hatt;
-pub use hatt::hatt;
+pub use anneal::*;
 mod topphatt;
 pub use topphatt::*;